refactoring: remove unused imports, clear useless space, remove types folder
