--- client/src/components/UI/Slider.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Slider = () => {
-    return (
-        <div>
-            slider
-        </div>
-    );
-};
-
-export default Slider;
