+++ client/src/models/ITheme.ts
@@ -1,5 +1,5 @@
 import React from "react";
-import {IIcon} from "../types/IIcon";
+import {IIcon} from "./IIcon";
 
 export interface ITheme {
     id: number,
