+++ client/src/pages/Games.tsx
@@ -5,7 +5,6 @@ import {useAppDispatch, useAppSelector} from "../hooks/redux";
 
 const Games = () => {
     const {games, isLoading, error} = useAppSelector(state => state.game)
-    const dispatch = useAppDispatch()
 
     return (
         <MainLayout>
